refactor(control): split startup preconnect
